Extract DatabasePoolSetupTrait
